requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15.1,<0.16
Pillow >=9.3
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.13.0
celery>=5.2.3,<6.0
redis==3.4.1
requests==2.28.1
django-filter==22.1
django-redis==5.2.0
gunicorn==20.1.0
django-debug-toolbar==3.8.1 #https://django-debug-toolbar.readthedocs.io/en/latest/
XlsxWriter #https://xlsxwriter.readthedocs.io/
pycryptodome #https://github.com/Legrandin/pycryptodome/
pendulum==2.1.2 #https://github.com/sdispater/pendulum
PyJWT==2.6.0 #https://pypi.org/project/PyJWT/
django-celery-beat==2.4.0 #http://github.com/celery/django-celery-beat
django-cloudinary-storage>=0.3.0 #https://github.com/klis87/django-cloudinary-storage
cloudinary>=1.34.0 #https://github.com/cloudinary/pycloudinary
django-dbbackup>=4.0.2
pandas>=2.1.0 #https://github.com/pandas-dev/pandas
odfpy>=1.4.1
